backend/requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
dill==0.4.0
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
numpy==2.3.0
packaging==25.0
pbr==6.1.1
pillow==11.2.1
pymongo==4.13.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qiskit==2.0.3
qiskit-aer==0.12.0
rustworkx==0.16.0
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
stevedore==5.4.1
symengine==0.13.0
sympy==1.14.0
typing_extensions==4.14.0
Werkzeug==3.1.3
